# 2–3x speed-up: batch all s3_exists into parallel calls