--- a/lib/shared.bash
+++ b/lib/shared.bash
@@ -1,9 +1,61 @@
 #!/bin/bash
 
+declare -a s3_cache_found
+declare -a s3_cache_not_found
+
+s3_exists_preload() {
+  local bucket="$1"; shift
+  local keys=("$@")
+
+  echo >&2 "Searching $bucket"
+  declare -a pid_to_cache_key
+  declare -a pid_to_key
+  for key in "${keys[@]}"; do
+    cache_key="$bucket/$key"
+    s3_exists "$bucket" "$key" &
+    pid=$!
+    pid_to_cache_key[$pid]="$cache_key"
+    pid_to_key[$pid]="$key"
+  done
+  declare -a keys_found
+  declare -a keys_not_found
+  for pid in "${!pid_to_cache_key[@]}"; do
+    cache_key=${pid_to_cache_key[$pid]}
+    key=${pid_to_key[$pid]}
+    if wait "$pid"; then
+      s3_cache_found+=("$cache_key")
+      keys_found+=("$key")
+    else
+      s3_cache_not_found+=("$cache_key")
+      keys_not_found+=("$key")
+    fi
+  done
+  echo >&2 "Found:"
+  for key in "${keys_found[@]}"; do
+    echo >&2 "- $key"
+  done
+  echo >&2 "Not found:"
+  for key in "${keys_not_found[@]}"; do
+    echo >&2 "- $key"
+  done
+}
+
 s3_exists() {
   local bucket="$1"
   local key="$2"
   local aws_s3_args=("--region=$AWS_DEFAULT_REGION")
+
+  cache_key="$bucket/$key"
+  for found in "${s3_cache_found[@]-}"; do
+    if [[ $found == "$cache_key" ]]; then
+      return 0
+    fi
+  done
+  for not_found in "${s3_cache_not_found[@]-}"; do
+    if [[ $not_found == "$cache_key" ]]; then
+      return 1
+    fi
+  done
 
   # capture just stderr
   output=$(aws s3api head-object "${aws_s3_args[@]}" --bucket "$bucket" --key "$key" 2>&1 >/dev/null)