Ten S3 object existence checks are parallelised instead of running sequentially.
21 seconds reduced to 7 seconds in my local environment.
Caveat: my round-trip latency is higher than within AWS.

This pull request introduces a `s3_exists_preload` function, which takes a list of all S3 keys and checks for existence in parallel. The results are cached for use in the regular `s3_exists` function.

- [ ] Somehow get the brittle BATS test suite passing.

Before:

    ~~~ Downloading secrets from :s3: buildkite-sandbox-pda-test
    Checking bar/private_ssh_key
    Checking bar/id_rsa_github
    Checking private_ssh_key
    Checking id_rsa_github
    Checking env
    Checking environment
    Downloading env file from environment
    Evaluating 13 bytes of env
    Checking bar/env
    Downloading env file from bar/env
    Evaluating 11 bytes of env
    Checking bar/environment
    ./hooks/environment  4.82s user 1.25s system 28% cpu 21.465 total

After:

    ~~~ Downloading secrets from :s3: buildkite-sandbox-pda-test
    Searching buildkite-sandbox-pda-test
    Found:
    - environment
    - bar/env
    Not found:
    - bar/private_ssh_key
    - bar/id_rsa_github
    - private_ssh_key
    - id_rsa_github
    - env
    - bar/environment
    - git-credentials
    - bar/git-credentials
    Downloading env file from environment
    Evaluating 13 bytes of env
    Downloading env file from bar/env
    Evaluating 11 bytes of env
    ./hooks/environment  7.61s user 2.10s system 133% cpu 7.275 total