@pda 2020-10-14
Here's a before/after test from an EC2 instance, showing more moderate speed-up:

```
$ git checkout master
Switched to branch 'master'
Your branch is up-to-date with 'origin/master'.

$ time ./hooks/environment
~~~ Downloading secrets from :s3: buildkite-sandbox-pda-test
Checking bar/private_ssh_key
Checking bar/id_rsa_github
Checking private_ssh_key
Checking id_rsa_github
Checking env
Checking environment
Downloading env file from environment
Evaluating 13 bytes of env
Checking bar/env
Downloading env file from bar/env
Evaluating 11 bytes of env
Checking bar/environment

real    0m6.812s
user    0m5.800s
sys     0m0.644s

$ git checkout parallel
Switched to branch 'parallel'
Your branch is up-to-date with 'origin/parallel'.

$ time ./hooks/environment
~~~ Downloading secrets from :s3: buildkite-sandbox-pda-test
Searching buildkite-sandbox-pda-test
Found:
- environment
- bar/env
Not found:
- bar/private_ssh_key
- bar/id_rsa_github
- private_ssh_key
- id_rsa_github
- env
- bar/environment
- git-credentials
- bar/git-credentials
Downloading env file from environment
Evaluating 13 bytes of env
Downloading env file from bar/env
Evaluating 11 bytes of env

real    0m4.450s
user    0m6.436s
sys     0m0.680s
```

In this test I'm using IAM credentials in my environment.
In reality, each call has some overhead acquiring session credentials from the EC2 instance metadata service.
So, the real-world improvement is probably a bit better than this.