@pda 2020-11-24
Closed, #37 merged instead.